Restructure unittests
