+++ tests/test_ir/test_data/compressed_benchmark.clp.zst
