+++ tests/test_ir/test_data/rand_hadoop_log.clp
