+++ tests/test_ir/test_log_event.py
@@ -3,9 +3,9 @@ from datetime import tzinfo
 from typing import Optional
 
 import dateutil.tz
+from test_ir.test_utils import TestCLPBase
 
 from clp_ffi_py import LogEvent, Metadata
-from tests.test_utils import TestCLPBase
 
 
 class TestCaseLogEvent(TestCLPBase):
